+++ src/commercetools/platform/client/search/by_project_key_product_projections_search_request_builder.py
@@ -15,6 +15,7 @@ if typing.TYPE_CHECKING:
 
 
 class ByProjectKeyProductProjectionsSearchRequestBuilder:
+
     _client: "BaseClient"
     _project_key: str
 
@@ -54,7 +55,7 @@ class ByProjectKeyProductProjectionsSearchRequestBuilder:
         self,
         *,
         fuzzy: bool = None,
-        fuzzy_level: float = None,
+        fuzzy_level: int = None,
         mark_matching_variants: bool = None,
         filter: typing.List["str"] = None,
         filter_facets: typing.List["str"] = None,
